Serializer

+++ db/migrate/20231005193634_create_doctors.rb
@@ -3,8 +3,8 @@ class CreateDoctors < ActiveRecord::Migration[7.0]
     create_table :doctors do |t|
       t.string :name
       t.string :specialization
-      t.decimal :consultation_fee
-      t.decimal :prescription_fee
+      t.integer :consultation_fee
+      t.integer :prescription_fee
 
       t.timestamps
     end
